+++ changelog.md
@@ -5,13 +5,26 @@ Unreleased
 
 New features:
 
+Updated features:
+
+Removed features:
+
+Internal improvements:
+
+Testing improvements:
+
+Others:
+
+v3.1.0 (2022-16-01)
+======
+
+New features:
+
 * mask_tokens config (default: true) to mask tokens in logs
 
 Updated features:
 
 * Log before and after sending request
-
-Removed features:
 
 Internal improvements:
 
@@ -20,9 +33,6 @@ Internal improvements:
 Testing improvements:
 
 * Run Travis on latest rubies and fix config
-
-Others:
-
 
 v3.0.0 (2017-03-17)
 ======